XLConnect-deprecated {XLConnect}	R Documentation

Deprecated functions in package XLConnect

Description

These functions are provided for compatibility with older versions of XLConnect only, and will be defunct in a later release.

Details

The following functions are deprecated and will be made defunct. Use the replacements as indicated.

getReferenceCoordinates: getReferenceCoordinatesForName

[Package XLConnect version 0.2-13 Index]