XLConnect-package {XLConnect}	R Documentation

Excel Connector for R

Description

Provides comprehensive functionality to read, write and format Excel data.

Details

Package: 	 XLConnect

Type: 	 Package

Version: 	 0.2-13

Date: 	 2017-05-10

URL: 	 http://www.mirai-solutions.com

http://miraisolutions.wordpress.com

SystemRequirements: 	 java (>= 1.6)

Depends: 	 R (>= 2.10.0), XLConnectJars (== 0.2-12)

Imports: 	 methods, rJava, utils

Suggests: 	 RUnit, lattice, ggplot2 (>= 0.9.3), zoo

License: 	 GPL-3

Copyright: 	 See file COPYRIGHTS

LazyData: 	 yes

For an overview over the package please refer to the available demos:
demo(package = "XLConnect")

Author(s)

Mirai Solutions GmbH, xlconnect@mirai-solutions.com

References

Mirai Solutions GmbH: http://www.mirai-solutions.com
Mirai Solutions on GitHub: https://github.com/miraisolutions
Mirai Solutions Blog: http://miraisolutions.wordpress.com
Apache POI: http://poi.apache.org

Examples

# Load workbook; create if not existing
wb <- loadWorkbook("XLConnect.xlsx", create = TRUE)

# Create a worksheet
createSheet(wb, name = "mtcars")

# Create a name reference
createName(wb, name = "mtcars", formula = "mtcars!$C$5")

# Write built-in data.frame 'mtcars' to the specified named region
writeNamedRegion(wb, mtcars, name = "mtcars")

# Save workbook
saveWorkbook(wb)

[Package XLConnect version 0.2-13 Index]