aref {XLConnect}	R Documentation

Constructing Excel area references

Description

Constructs an Excel area reference

Usage

aref(topLeft, dimension)

Arguments

topLeft
Top left corner. Either a character specifying a cell reference in the form "A1" or a numeric vector of length two specifying the corresponding coordinates.

dimension
Dimensions (numeric) of a 2-dimensional object (mostly a data.frame or a matrix)

Value

Returns the area reference (character) for the specified top left cell and dimension.

Author(s)

Martin Studer
Mirai Solutions GmbH http://www.mirai-solutions.com

See Also

aref2idx, idx2aref, idx2cref, col2idx, idx2col

Examples

aref("A1", dim(mtcars))
aref(c(1, 1), dim(mtcars))

[Package XLConnect version 0.2-13 Index]