aref2idx {XLConnect}	R Documentation

Converting Excel cell references to row and column based cell references

Description

Converts Excel cell references to row and column based cell references

Usage

aref2idx(x)

Arguments

x
Character vector of Excel cell references (e.g. "A1:B6", "B6:C17", ...)

Value

Returns a numeric matrix with four columns and as many rows as cell references that have been provided. The first two columns represent the coordinates of the top left corner (row, column) and the third and fourth columns represent the bottom right corner of the referenced area.

Author(s)

Martin Studer
Mirai Solutions GmbH http://www.mirai-solutions.com

See Also

idx2aref, aref, cref2idx, idx2cref, col2idx, idx2col

Examples

aref2idx(c("A1:B6", "B6:C17"))

[Package XLConnect version 0.2-13 Index]