clearSheet-methods {XLConnect}	R Documentation

Clearing worksheets in a workbook

Description

Clears worksheets with specified names or indices in a workbook.

Usage

  ## S4 method for signature 'workbook,numeric'
clearSheet(object, sheet)
  ## S4 method for signature 'workbook,character'
clearSheet(object, sheet)

Arguments

object
The workbook to use

sheet
The name or the index of the worksheet to clear

Details

Clearing a worksheet means to clear all the cells in that worksheet. Consequently, the saved workbook should be smaller in size. Clearing a worksheet can be useful if data sets in a worksheet need to be replaced, i.e. data are first read, modified in R and finally written back to the worksheet. Without clearing the worksheet first, (parts of) the original data may still be visible if they occupied a larger range of the worksheet.

Author(s)

Nicola Lambiase
Mirai Solutions GmbH http://www.mirai-solutions.com

See Also

workbook, clearNamedRegion, clearRange, clearRangeFromReference

Examples

# mtcars xlsx file from demoFiles subfolder of 
# package XLConnect
demoExcelFile <- system.file("demoFiles/mtcars.xlsx", 
                             package = "XLConnect")

# Load workbook
wb <- loadWorkbook(demoExcelFile)

# Clear worksheets named 'mtcars' and 'mtcars2'
clearSheet(wb, sheet = c("mtcars", "mtcars2"))

# Clear 3rd worksheet
clearSheet(wb, sheet = 3) 

[Package XLConnect version 0.2-13 Index]