cloneSheet-methods {XLConnect}	R Documentation

Cloning/copying worksheets

Description

Clones (copies) a worksheet in a workbook.

Usage

## S4 method for signature 'workbook,numeric'
cloneSheet(object,sheet,name)
## S4 method for signature 'workbook,character'
cloneSheet(object,sheet,name)

Arguments

object
The workbook to use

sheet
The name or index of the worksheet to clone

name
The name to assign to the cloned worksheet. Throws an exception if the name to assign is the name of an already existing worksheet.

Author(s)

Martin Studer
Mirai Solutions GmbH http://www.mirai-solutions.com

See Also

workbook, createSheet, removeSheet, renameSheet, getSheets, existsSheet

Examples

# mtcars xlsx file from demoFiles subfolder of package XLConnect
demoExcelFile <- system.file("demoFiles/mtcars.xlsx", package = "XLConnect")

# Load workbook
wb <- loadWorkbook(demoExcelFile)

# Clone the 'mtcars' worksheet and assign it the name 'mtcars cloned'
cloneSheet(wb, sheet = "mtcars", name = "mtcars cloned")

[Package XLConnect version 0.2-13 Index]