col2idx {XLConnect}	R Documentation

Converting Excel column names to indices

Description

Converts Excel column names to indices.

Usage

col2idx(x)

Arguments

x
Character vector of Excel column names (e.g. "A", "AF", ...)

Value

Returns a vector of integers representing the corresponding column indices. Note that passing invalid column name references may result in an arbitrary number.

Author(s)

Martin Studer
Mirai Solutions GmbH http://www.mirai-solutions.com

See Also

idx2col, cref2idx, idx2cref, idx2aref, aref2idx, aref

Examples

col2idx(c("A", "BTG"))

[Package XLConnect version 0.2-13 Index]