createName-methods {XLConnect}	R Documentation

Creating names in a workbook

Description

Creates a name for a specified formula in a workbook.

Usage

## S4 method for signature 'workbook'
createName(object, name, formula, overwrite)

Arguments

object
The workbook to use

name
The name's name to create

formula
Excel formula specifying the name

overwrite
If a name with the same name already exists and overwrite = TRUE, then this name is removed first before the new one is created. If a name already exists and overwrite = FALSE, then an exception is thrown. The default value for overwrite is FALSE.

Details

Creates a name named name for the specified formula.

The formula should be specified as you would type it in Excel. Make sure that the worksheets, functions, ... exist that you are referring to in the formula.

The name, formula and overwrite arguments are vectorized such that multiple names can be created in one method call.

Author(s)

Martin Studer
Mirai Solutions GmbH http://www.mirai-solutions.com

References

What are named regions/ranges?
http://www.officearticles.com/excel/named_ranges_in_microsoft_excel.htm
How to create named regions/ranges?
http://www.youtube.com/watch?v=iAE9a0uRtpM

See Also

workbook, removeName, existsName, getDefinedNames,
readNamedRegion, writeNamedRegion

Examples

# Load workbook (create if not existing)
wb <- loadWorkbook("createName.xlsx", create = TRUE)

# Create a worksheet named 'mtcars'
createSheet(wb, name = "mtcars")

# Create a named region called 'mtcars' on the sheet called 'mtcars'
createName(wb, name = "mtcars", formula = "mtcars!$A$1")

# Write built-in data set 'mtcars' to the above defined named region
writeNamedRegion(wb, mtcars, name = "mtcars")

# Save workbook
saveWorkbook(wb)

[Package XLConnect version 0.2-13 Index]