cref2idx {XLConnect}	R Documentation

Converting Excel cell references to indices

Description

Converts Excel cell references to row & column indices

Usage

cref2idx(x)

Arguments

x
Character vector of Excel cell references (e.g. "$A$20", "B18", ...)

Value

Returns a numeric matrix with two columns and as many rows as cell references that have been provided. The first column represents the row indices and the second column represents the column indices.

Author(s)

Martin Studer
Mirai Solutions GmbH http://www.mirai-solutions.com

See Also

idx2cref, col2idx, idx2col, idx2aref, aref2idx, aref

Examples

cref2idx(c("$A$20", "B18"))

[Package XLConnect version 0.2-13 Index]