existsSheet-methods {XLConnect}	R Documentation

Checking for existence of worksheets in a workbook

Description

Checks the existence of a worksheet in a workbook.

Usage

## S4 method for signature 'workbook'
existsSheet(object,name)

Arguments

object
The workbook to use

name
The sheet name to check for

Details

Checks if the specified worksheet exists. Returns TRUE if it exists, otherwise FALSE. The name argument is vectorized which allows to check for existence of multiple worksheets with one call.

Author(s)

Martin Studer
Mirai Solutions GmbH http://www.mirai-solutions.com

See Also

workbook, createSheet, removeSheet, renameSheet, getSheets, cloneSheet

Examples

# mtcars xlsx file from demoFiles subfolder of package XLConnect
demoExcelFile <- system.file("demoFiles/mtcars.xlsx", package = "XLConnect")

# Load workbook
wb <- loadWorkbook(demoExcelFile)

# Check for existence of a worksheet called 'mtcars'
existsSheet(wb, "mtcars")

[Package XLConnect version 0.2-13 Index]