getCellFormula-methods {XLConnect}	R Documentation

Retrieving formula definitions from cells

Description

Retrieves a cell formula from a workbook.

Usage

## S4 method for signature 'workbook,character'
getCellFormula(object,sheet,row,col)
## S4 method for signature 'workbook,numeric'
getCellFormula(object,sheet,row,col)

Arguments

object
The workbook to use

sheet
The name or index of the worksheet containing the cell

row
The one-based row index of the cell to query

col
The one-based column index of the cell to query

Details

Retrieves the formula of the specified cell as a character, without the initial = character displayed in Excel. Raises an error if the specified cell is not a formula cell.

Author(s)

Thomas Themel
Mirai Solutions GmbH http://www.mirai-solutions.com

See Also

workbook, setCellFormula

Examples

# Load workbook (create if not existing)
wb <- loadWorkbook("cellFormula.xlsx", create = TRUE)

createSheet(wb, "Formula")

# Assign a formula to A1
setCellFormula(wb, "Formula", 1, 1, "SUM($B$1:$B$29)")

# Returns the formula for Sheet1!A1
getCellFormula(wb, "Formula", 1, 1)
# The same with a numeric sheet index
getCellFormula(wb, 1, 1, 1)

[Package XLConnect version 0.2-13 Index]