getForceFormulaRecalculation-methods {XLConnect}	R Documentation

Querying the coordinates of the range reference by an Excel name

Description

Queries the "force formula recalculation" flag on an Excel worksheet.

Usage

## S4 method for signature 'workbook,character'
getForceFormulaRecalculation(object,sheet)
## S4 method for signature 'workbook,numeric'
getForceFormulaRecalculation(object,sheet)

Arguments

object
The workbook to use

sheet
The name or index of the sheet to query. This argument is vectorized such that multiple sheets can be queried with one method call. If sheet = "*", the flag is queried for all sheets in the workbook (in the order as returned by getSheets).

Author(s)

Thomas Themel
Mirai Solutions GmbH http://www.mirai-solutions.com

See Also

workbook, getSheets, setForceFormulaRecalculation

Examples

# mtcars xlsx file from demoFiles subfolder of package XLConnect
demoExcelFile <- system.file("demoFiles/mtcars.xlsx", package = "XLConnect")

# Load workbook
wb <- loadWorkbook(demoExcelFile)

# Ask whether Excel will automatically recalculate formulas on sheet mtcars
print(getForceFormulaRecalculation(wb, sheet = "mtcars"))

[Package XLConnect version 0.2-13 Index]