getReferenceCoordinatesForName-methods {XLConnect}	R Documentation

Querying the coordinates of the range reference by an Excel name

Description

Queries the coordinates of an Excel name in a workbook.

Usage

## S4 method for signature 'workbook'
getReferenceCoordinatesForName(object,name)

Arguments

object
The workbook to use

name
The name to query. This argument is vectorized such that multiple names can be queried with one method call.

Author(s)

Thomas Themel
Mirai Solutions GmbH http://www.mirai-solutions.com

See Also

workbook, createName, existsName, removeName, getReferenceFormula, getReferenceCoordinatesForTable

Examples

# mtcars xlsx file from demoFiles subfolder of package XLConnect
demoExcelFile <- system.file("demoFiles/mtcars.xlsx", package = "XLConnect")

# Load workbook
wb <- loadWorkbook(demoExcelFile)

# Query reference coordinate for name 'mtcars'
print(getReferenceCoordinatesForName(wb, name = "mtcars"))

[Package XLConnect version 0.2-13 Index]