getReferenceFormula-methods {XLConnect}	R Documentation

Querying reference formulas of Excel names

Description

Queries the reference formula of an Excel name in a workbook.

Usage

## S4 method for signature 'workbook'
getReferenceFormula(object,name)

Arguments

object
The workbook to use

name
The name to query. This argument is vectorized such that multiple names can be queried with one method call.

Author(s)

Martin Studer
Mirai Solutions GmbH http://www.mirai-solutions.com

See Also

workbook, createName, existsName, removeName

Examples

# mtcars xlsx file from demoFiles subfolder of package XLConnect
demoExcelFile <- system.file("demoFiles/mtcars.xlsx", package = "XLConnect")

# Load workbook
wb <- loadWorkbook(demoExcelFile)

# Query reference formula for name 'mtcars'
print(getReferenceFormula(wb, name = "mtcars"))

[Package XLConnect version 0.2-13 Index]