getSheetPos-methods {XLConnect}	R Documentation

Querying worksheet position

Description

Queries the position of a worksheet in a workbook.

Usage

## S4 method for signature 'workbook,character'
getSheetPos(object,sheet)

Arguments

object
The workbook to use

sheet
The name of the worksheet (character) to query. This argument is vectorized such that multiple worksheets can be queried with one method call.

Value

Returns the position index of the corresponding worksheet. Note that querying a non-existing worksheet results in a 0 index and does not throw an exception!

Author(s)

Martin Studer
Mirai Solutions GmbH http://www.mirai-solutions.com

See Also

workbook, setSheetPos, getSheets

Examples

# mtcars xlsx file from demoFiles subfolder of package XLConnect
demoExcelFile <- system.file("demoFiles/mtcars.xlsx", package = "XLConnect")

# Load workbook
wb <- loadWorkbook(demoExcelFile)

# Query worksheet positions for the worksheets 'mtcars2', 'mtcars3',
# 'mtcars' and 'NotThere' (which actually does not exist)
print(getSheetPos(wb, sheet = c("mtcars2", "mtcars3", "mtcars", "NotThere")))

[Package XLConnect version 0.2-13 Index]