getTables-methods {XLConnect}	R Documentation

Querying available Excel tables in a workbook

Description

Queries the available Excel tables on the specified worksheet.

Usage

## S4 method for signature 'workbook,numeric'
getTables(object,sheet,simplify)
## S4 method for signature 'workbook,character'
getTables(object,sheet,simplify)

Arguments

object
The workbook to use

sheet
Index (integer) or name (character) of worksheet to query

simplify
logical specifying if the result should be simplified (defaults to TRUE). See details.

Details

Since this is a vectorized function (multiple sheets can be specified) the result is a named list (one component per sheet) if no simplification is applied. In cases where only one sheet is queried and simplify = TRUE (default) the result is simplified to a vector.

Author(s)

Martin Studer
Mirai Solutions GmbH http://www.mirai-solutions.com

See Also

workbook, getSheets, readTable

Examples

# mtcars xlsx file from demoFiles subfolder of package XLConnect
demoExcelFile <- system.file("demoFiles/mtcars.xlsx", package = "XLConnect")

# Load workbook
wb <- loadWorkbook(demoExcelFile)

# Query available tables (table names) on sheet 'mtcars_table'
tables <- getTables(wb, sheet = "mtcars_table")

# ... or via sheet index
tables <- getTables(wb, sheet = 4)

[Package XLConnect version 0.2-13 Index]