idx2aref {XLConnect}	R Documentation

Converting row and column based area references to Excel area references

Description

Converts row & column based area references to Excel area references

Usage

idx2aref(x)

Arguments

x
Numeric (integer) matrix or vector of indices. If a matrix is provided it should have four columns with the first two columns representing the top left corner (row and column indices) and the third & fourth column representing the bottom right corner. If a vector is provided it will be converted to a matrix by filling the vector into a 4-column matrix by row.

Value

Returns a character vector of corresponding Excel area references.

Author(s)

Martin Studer
Mirai Solutions GmbH http://www.mirai-solutions.com

See Also

aref2idx, aref, idx2cref, cref2idx, idx2col, col2idx

Examples

idx2aref(c(1, 1, 5, 4))

[Package XLConnect version 0.2-13 Index]