idx2cref {XLConnect}	R Documentation

Converting indices to Excel cell references

Description

Converts row & column indices to Excel cell references

Usage

idx2cref(x, absRow = TRUE, absCol = TRUE)

Arguments

x
Numeric (integer) matrix or vector of indices. If a matrix is provided it should have two columns with the first column representing the row indices and the second column representing the column indices (i.e. each row represents a index-based cell reference). If a vector is provided it will be converted to a matrix by filling the vector into a 2-column matrix by row.

absRow
Boolean determining if the row index should be considered absolute. If TRUE (default), this will result in a '$'-prefixed row identifier.

absCol
Boolean determining if the column index should be considered absolute. If TRUE (default), this will result in a '$'-prefixed column identifier.

Value

Returns a character vector of corresponding Excel cell references.

Author(s)

Martin Studer
Mirai Solutions GmbH http://www.mirai-solutions.com

See Also

cref2idx, idx2col, col2idx, idx2aref, aref2idx, aref

Examples

idx2cref(c(5, 8, 14, 38))

[Package XLConnect version 0.2-13 Index]