jTryCatch {XLConnect}	R Documentation

Standard Java exception handling for XLConnect

Description

Catches Java exceptions to create nicer error messages.

Usage

jTryCatch(...)

Arguments

...
Arguments to tryCatch - an expression and optionally the argument finally to specify an expression to be evaluated before returning or exiting.

Details

This function specifies a Java Throwable error handler that catches any Java exceptions. Caught exceptions are post-processed and re-thrown using stop.

Author(s)

Martin Studer
Mirai Solutions GmbH http://www.mirai-solutions.com

See Also

tryCatch, stop

Examples

try(XLConnect:::jTryCatch(J("java.lang.Integer")$valueOf("not an integer")))

[Package XLConnect version 0.2-13 Index]