normalizeDataframe {XLConnect}	R Documentation

Data frame Normalization for Unit Tests

Description

Normalizes a data.frame with respect to column types. This is necessary for comparing data.frames written to Excel and then read from Excel. Excel does not know e.g. factor types. Factor variables in fact are written to Excel as ordinary strings. Therefore, when read back in to R, they are treated as character variables.

Usage

normalizeDataframe(df, replaceInf)

Arguments

df
data.frame to normalize

replaceInf
logical specifying if infinites should be replaced with NA; this is used because Excel does not support infinites

Value

Returns a normalized data.frame

Author(s)

Martin Studer
Mirai Solutions GmbH http://www.mirai-solutions.com

[Package XLConnect version 0.2-13 Index]