readWorksheetFromFile {XLConnect}	R Documentation

Reading data from worksheets in an Excel file (wrapper function)

Description

Reads data from worksheets in an Excel file.

Usage

readWorksheetFromFile(file, ...)

Arguments

file
The path name of the file to read from.

...
Arguments passed to readWorksheet

Details

See readWorksheet for more information.

Author(s)

Thomas Themel
Mirai Solutions GmbH http://www.mirai-solutions.com

See Also

readWorksheet, readNamedRegionFromFile, writeWorksheetToFile,
writeNamedRegionToFile, onErrorCell

Examples

# multiregion xlsx file from demoFiles subfolder of 
# package XLConnect
demoExcelFile <- system.file("demoFiles/multiregion.xlsx", 
                             package = "XLConnect")

# Read single area from first sheet of existing file,
# "B2:C3" in Excel speak
df.one <- readWorksheetFromFile(demoExcelFile, sheet = 1, 
                                header = FALSE, startCol = 2, 
                                startRow = 2, endCol = 3, 
                                endRow = 3)

# Read three data sets in one from known positions
dflist <- readWorksheetFromFile(demoExcelFile,
                                sheet = c("FirstSheet", 
                                          "FirstSheet", 
                                          "SecondSheet"),
                                header = TRUE, 
                                startRow = c(2,2,3), 
                                startCol = c(2,5,2),
                                endCol = c(5,8,6), 
                                endRow = c(9,15,153))

[Package XLConnect version 0.2-13 Index]