setActiveSheet-methods {XLConnect}	R Documentation

Setting the active worksheet in a workbook

Description

Sets the active worksheet of a workbook.

Usage

## S4 method for signature 'workbook,character'
setActiveSheet(object,sheet)
## S4 method for signature 'workbook,numeric'
setActiveSheet(object,sheet)

Arguments

object
The workbook to use

sheet
The name or index of the sheet to activate

Note

The active worksheet of a workbook is the worksheet that is displayed when the corresponding Excel file is opened.

Author(s)

Martin Studer
Mirai Solutions GmbH http://www.mirai-solutions.com

See Also

workbook, createSheet, removeSheet, renameSheet, existsSheet, getSheets

Examples

# mtcars xlsx file from demoFiles subfolder of package XLConnect
mtcarsFile <- system.file("demoFiles/mtcars.xlsx", package = "XLConnect")

# Load workbook
wb <- loadWorkbook(mtcarsFile)

# Sets the active sheet to the sheet 'mtcars3'
setActiveSheet(wb, sheet = "mtcars3")

[Package XLConnect version 0.2-13 Index]