setDataFormat-methods {XLConnect}	R Documentation

Specifying custom data formats for cell styles

Description

Specifies a custom data format for a cellstyle.

Usage

## S4 method for signature 'cellstyle'
setDataFormat(object,format)

Arguments

object
The cellstyle to use

format
A data format string

Details

Specifies the data format to be used by the corresponding cellstyle. Data formats are specified the standard Excel way. Refer to the Excel help or to the link below for more information.

Author(s)

Martin Studer
Mirai Solutions GmbH http://www.mirai-solutions.com

References

Excel custom data formats:
http://www.ozgrid.com/Excel/CustomFormats.htm

See Also

workbook, cellstyle, setCellStyle, setStyleAction

Examples

# Load workbook (create if not existing)
wb <- loadWorkbook("setDataFormat.xlsx", create = TRUE)

# Create a worksheet
createSheet(wb, name = "cellstyles")

# Create a dummy data set with the current date/time (as POSIXct)
now <- data.frame(Now = Sys.time())

# Write the value to the 'cellstyles' worksheet in the top left 
# corner (cell A1)
writeWorksheet(wb, now, sheet = "cellstyles", startRow = 1, 
               startCol = 1, header = FALSE)

# Create a custom anonymous cell style
cs <- createCellStyle(wb)

# Specify a custom data format
setDataFormat(cs, format = "dddd d-m-yyyy h:mm AM/PM")

# Set the cell style created above for the top left cell (A1) in 
# the 'cellstyles' worksheet
setCellStyle(wb, sheet = "cellstyles", row = 1, col = 1, cellstyle = cs)

# Set column width to display whole time/date string
setColumnWidth(wb, sheet = "cellstyles", column = 1, width = 6000)

# Save the workbook
saveWorkbook(wb)

[Package XLConnect version 0.2-13 Index]