setRowHeight-methods {XLConnect}	R Documentation

Setting the height of a row in a worksheet

Description

Sets the height of a row in a worksheet.

Usage

## S4 method for signature 'workbook,character'
setRowHeight(object,sheet,row,height)
## S4 method for signature 'workbook,numeric'
setRowHeight(object,sheet,row,height)

Arguments

object
The workbook to use

sheet
The name or index of the sheet to edit

row
The index of the row to resize

height
The height in points. If height < 0 (default: -1), the row will be sized to the sheet's default row height.

Details

Note that the arguments sheet, row and height are vectorized. As such the row height of multiple rows (potentially on different worksheets) can be set with one method call.

Author(s)

Martin Studer
Mirai Solutions GmbH http://www.mirai-solutions.com

See Also

workbook, setColumnWidth

Examples

# mtcars xlsx file from demoFiles subfolder of package XLConnect
mtcarsFile <- system.file("demoFiles/mtcars.xlsx", package = "XLConnect")

# Load workbook
wb <- loadWorkbook(mtcarsFile)

# Sets the row height of the 1st row on sheet 'mtcars'
# to 20 points
setRowHeight(wb, sheet = "mtcars", row = 1, height = 20)

[Package XLConnect version 0.2-13 Index]