setSheetPos-methods {XLConnect}	R Documentation

Setting worksheet position

Description

Sets the position of a worksheets in a workbook.

Usage

## S4 method for signature 'workbook,character,numeric'
setSheetPos(object,sheet,pos)

Arguments

object
The workbook to use

sheet
The name of the worksheet (character) whose position to set. This argument is vectorized such that the positions of multiple worksheets can be set with one method call.

pos
The position index to set for the corresponding sheet. If missing, sheets will be positioned in the order they are specified in the argument sheet.

Details

It is important to note that the worksheet positions will be applied one after the other in the order they have been specified.

Author(s)

Martin Studer
Mirai Solutions GmbH http://www.mirai-solutions.com

See Also

workbook, getSheetPos, getSheets

Examples

# mtcars xlsx file from demoFiles subfolder of package XLConnect
demoExcelFile <- system.file("demoFiles/mtcars.xlsx", package = "XLConnect")

# Load workbook
wb <- loadWorkbook(demoExcelFile)

# Move the 'mtcars3' worksheet to the front
setSheetPos(wb, sheet = "mtcars3", pos = 1)

[Package XLConnect version 0.2-13 Index]