setWrapText-methods {XLConnect}	R Documentation

Specifying text wrapping behaviour

Description

Specifies if text should be wrapped in a cell.

Usage

## S4 method for signature 'cellstyle'
setWrapText(object,wrap)

Arguments

object
The cellstyle to manipulate

wrap
If wrap = TRUE, the text is wrapped if it exceeds the width of the cell - otherwise not.

Author(s)

Martin Studer
Mirai Solutions GmbH http://www.mirai-solutions.com

See Also

workbook, cellstyle, setCellStyle, setStyleAction

Examples

# Load workbook (create if not existing)
wb <- loadWorkbook("setWrapText.xlsx", create = TRUE)

# Create a worksheet
createSheet(wb, name = "cellstyles")

# Create a dummy data set with some long text
text <- data.frame(
     Text = "Some very very very very very very very long text")

# Write the value to the 'cellstyles' worksheet in the 
# top left corner (cell A1)
writeWorksheet(wb, text, sheet = "cellstyles", startRow = 1, 
               startCol = 1, header = FALSE)

# Create a custom anonymous cell style
cs <- createCellStyle(wb)

# Specify to wrap the text
setWrapText(cs, wrap = TRUE)

# Set the cell style created above for the top left cell (A1) 
# in the 'cellstyles' worksheet
setCellStyle(wb, sheet = "cellstyles", row = 1, col = 1, 
             cellstyle = cs)

# Save the workbook
saveWorkbook(wb)

[Package XLConnect version 0.2-13 Index]