summary-methods {XLConnect}	R Documentation

Summarizing workbook objects

Description

Outputs a workbook summary including the underlying Excel filename, contained worksheets, hidden sheets, very hidden sheets, defined names and the active sheet name.

Usage

## S4 method for signature 'workbook'
summary(object)

Arguments

object
The workbook to summarize

Author(s)

Martin Studer
Mirai Solutions GmbH http://www.mirai-solutions.com

See Also

workbook, getSheets, isSheetHidden, isSheetVeryHidden, getDefinedNames, getActiveSheetName

Examples

# Load existing demo Excel file 'mtcars.xlsx' from the XLConnect package
wb.mtcars <- loadWorkbook(system.file("demoFiles/mtcars.xlsx", 
                          package = "XLConnect"))

# Print a workbook summary
summary(wb.mtcars)

[Package XLConnect version 0.2-13 Index]