unmergeCells-methods {XLConnect}	R Documentation

Unmerging cells

Description

Unmerges cells in a worksheet.

Usage

## S4 method for signature 'workbook,character'
unmergeCells(object,sheet,reference)
## S4 method for signature 'workbook,numeric'
unmergeCells(object,sheet,reference)

Arguments

object
The workbook to use

sheet
The name or index of the sheet on which to unmerge cells

reference
A cell range specification (character) in the form 'A1:B8'. Note that the specification must exactly correspond to the range of the merged cells.

Author(s)

Martin Studer
Mirai Solutions GmbH http://www.mirai-solutions.com

See Also

workbook, mergeCells, idx2cref

Examples

# Load workbook (create if not existing)
wb <- loadWorkbook("unmergeCells.xlsx", create = TRUE)

# Create a worksheet called 'merge'
createSheet(wb, name = "merge")

# Merge the cells A1:B8 on the worksheet created above
mergeCells(wb, sheet = "merge", reference = "A1:B8")

# Unmerge the cells A1:B8
unmergeCells(wb, sheet = "merge", reference = "A1:B8")

[Package XLConnect version 0.2-13 Index]