with.workbook {XLConnect}	R Documentation

Evaluate an R expression in a workbook environment

Description

Evaluate an R expression in an environment constructed from the named regions of an Excel workbook.

Usage

## S3 method for class 'workbook'
with(data, expr, ...)

Arguments

data
A workbook object, as returned by loadWorkbook.

expr
expression to evaluate

...
Additional arguments passed to readNamedRegion

Details

This method will read all named regions from the workbook when creating the environment. Names in the workbook will be processed through make.names to obtain the variable names.

Changes to the variables representing named regions will not affect the workbook contents and need to be saved explicitly using writeNamedRegion and saveWorkbook. If the workbook contains names that do not map to R identifiers,

Author(s)

Martin Studer
Mirai Solutions GmbH http://www.mirai-solutions.com

See Also

getDefinedNames, readNamedRegion,

Examples

# multiregion xlsx file from demoFiles subfolder of package XLConnect
demoExcelFile <- system.file("demoFiles/multiregion.xlsx",
                             package = "XLConnect")

# load workbook
wb <- loadWorkbook(demoExcelFile)

# named regions: Calendar, IQ, Iris
print(getDefinedNames(wb))

# named regions as variables
with(wb, { 
  print(Calendar)
  summary(IQ)
  summary(Iris)
})

[Package XLConnect version 0.2-13 Index]