wrapList {XLConnect}	R Documentation

Wrapping of arguments in a list

Description

Wraps the specified argument in a list.

Usage

wrapList(x)

Arguments

x
Any object.

Details

Wraps an argument in a list if it not already is a list. Further extracts the 'jobj' slot from S4 classes (such as workbook and cellstyle).

Value

list which wraps the provided argument if it not already was a list. If the provided argument is a S4 class with a 'jobj' slot, the corresponding slot is extracted and wrapped in a list.

Author(s)

Martin Studer
Mirai Solutions GmbH http://www.mirai-solutions.com

See Also

xlcCall

[Package XLConnect version 0.2-13 Index]