writeNamedRegionToFile {XLConnect}	R Documentation

Writing named regions to an Excel file (wrapper function)

Description

Writes named regions to an Excel file.

Usage

writeNamedRegionToFile(file, data, name, formula=NA, ...,
styleAction = XLC$STYLE_ACTION.XLCONNECT, clearNamedRegions=FALSE)

Arguments

file
The path name of the file to write to

data
Data to write

name
Name of the named region to write to

formula
If formula is specified, each item defines the formula of the named region identified by the corresponding entry of name. Use this if you want to create the document from scratch instead of writing to a template!

...
Additional arguments passed to writeNamedRegion

styleAction
Style action to be used when writing the data.
The default is XLC$STYLE_ACTION.XLCONNECT. See setStyleAction for more information.

clearNamedRegions
TRUE to clear content of existing named regions before writing data

Author(s)

Thomas Themel
Mirai Solutions GmbH http://www.mirai-solutions.com

See Also

writeNamedRegion, writeWorksheetToFile, readNamedRegionFromFile,
readWorksheetFromFile

Examples

# multiregion xlsx file from demoFiles subfolder of package XLConnect
demoExcelFile <- system.file("demoFiles/multiregion.xlsx", 
                             package = "XLConnect")

template <- "template-ws.xlsx"
file.copy(demoExcelFile, template) 

# Write single data.frame to a named region in an existing file
writeNamedRegionToFile(template, name = "Iris", iris)

# Write to a new file, defining the sheet and named region as we write.
# Format according to XLConnect defaults
writeNamedRegionToFile("iris.xlsx", name = "Iris", data = iris,
                       formula = "IrisData!$C$4", 
                       styleAction = "XLCONNECT")

[Package XLConnect version 0.2-13 Index]