writeWorksheet-methods {XLConnect}	R Documentation

Writing data to worksheets

Description

Writes data to worksheets of a workbook.

Usage

## S4 method for signature 'workbook,ANY,character'
writeWorksheet(object,data,sheet,startRow,startCol,header,rownames)
## S4 method for signature 'workbook,ANY,numeric'
writeWorksheet(object,data,sheet,startRow,startCol,header,rownames)

Arguments

object
The workbook to write to

data
Data to write

sheet
The name or index of the sheet to write to

startRow
Index of the first row to write to. The default is startRow = 1.

startCol
Index of the first column to write to. The default is startCol = 1.

header
Specifies if the column names should be written. The default is TRUE.

rownames
Name (character) of column to use for the row names of the provided data object. If specified, the row names of the data object (data.frame) will be included as an additional column with the specified name. If rownames = NULL (default), no row names will be included. May also be a list in case multiple data objects are written in one call (see below).

Details

Writes data to the worksheet specified by sheet. data is assumed to be a data.frame and is coerced to one if this is not already the case. startRow and startCol define the top left corner of the data region to be written. Note that the arguments are vectorized and as such multiple data objects (data.frame's) can be written to different worksheets in one call. In this case data is assumed to be a list of data.frames.

Note

When writing data to Excel, writeWorksheet further applies cell styles to the cells as defined by the workbook's "style action" (see setStyleAction).

Author(s)

Martin Studer
Mirai Solutions GmbH http://www.mirai-solutions.com

See Also

workbook, writeNamedRegion, appendWorksheet, appendNamedRegion, readWorksheet, readNamedRegion, writeWorksheetToFile

Examples

# Load workbook (create if not existing)
wb <- loadWorkbook("writeWorksheet.xlsx", create = TRUE)

# Create a worksheet called 'CO2'
createSheet(wb, name = "CO2")

# Write built-in data set 'CO2' to the worksheet created above;
# offset from the top left corner and with default header = TRUE
writeWorksheet(wb, CO2, sheet = "CO2", startRow = 4, startCol = 2)

# Save workbook (this actually writes the file to disk)
saveWorkbook(wb)

[Package XLConnect version 0.2-13 Index]