xlcEdit {XLConnect}	R Documentation

Editing data sets in an Excel file editor

Description

Provides the capability to edit an object/data.frame in an Excel file editor. After editing, the object is restored in the R session with the corresponding changes.

Usage

xlcEdit(obj, pos = globalenv(), ext = ".xlsx")

Arguments

obj
Object (data.frame) to edit.

pos
Where to look for the object specified by obj. See pos argument of get for more information.

ext
Extension to use for the Excel file being created. Defaults to ".xlsx".

Details

This function uses xlcDump and xlcRestore to dump objects to and restore objects from Excel files. An OS command is invoked to open the temporary Excel file in the default editor. Changes to the file have to be saved in order for them to take effect in the restored object.

Value

Invisibly returns the value of the xlcRestore operation.

Note

This function only works under Windows and MacOS with a corresponding Excel file editor, e.g. MS Excel or LibreOffice. Attempts to use this function under another OS will result in an error being thrown.

Author(s)

Martin Studer
Mirai Solutions GmbH http://www.mirai-solutions.com

See Also

xlcDump, xlcRestore, writeNamedRegion, writeWorksheet, writeNamedRegionToFile, writeWorksheetToFile

Examples

## Not run: 
myObj = mtcars
xlcEdit(myObj)

## End(Not run)

[Package XLConnect version 0.2-13 Index]