xlcFreeMemory {XLConnect}	R Documentation

Freeing Java Virtual Machine memory

Description

Frees Java Virtual Machine (JVM) memory.

Usage

xlcFreeMemory(...)

Arguments

...
Further arguments to be passed to R's garbage collector (gc).

Details

This function uses Java's Runtime class to run the garbage collector.
Java memory is freed by first running R's garbage collector (gc) and then Java's garbage collector. This sequence is important as R's gc may release objects which in turn allows Java's garbage collector to release some objects.

Note, in general there should be no need to make active use of this with XLConnect. Both R and Java automatically perform garbage collection at times. However, this function might be useful to reclaim memory after removing a large data object that has been written/read to/from Excel.

Author(s)

Martin Studer
Mirai Solutions GmbH http://www.mirai-solutions.com

See Also

xlcMemoryReport, gc

Examples

xlcFreeMemory()

[Package XLConnect version 0.2-13 Index]